kind of working version of network likes. still needs tests, better documentation, and a smarter way to make sure we get num_candidates, since theyre getting filtered out
